check if filenames are present in metadata
Added checks to make sure the name of each uploaded file is present in the
metadata JSON. If not, it returns a 400 error with the name of a missing
filename.

It does this before starting to write the zip to make sure a zip isn't aborted
halfway through writing it.