fix metadata truncation bug
Fixed a bug where metadata would get truncated if it was too long.
Apparently this was caused by the adm-zip library's addFile() function,
which I used to write the metadata string directly into the zip.
Instead we now write the metadata file to the temp directory first, then
zip it. This seemed to resolve the issue. Also I switched from
extractAllToAsync to extractAllTo in extractZip because it seems to have
fewer bugs.